create task

+++ src/main/java/cnpm/doan/repository/TaskRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface TaskRepository extends JpaRepository<Task, Integer> {
-    List<Task> findAllByProjectId(long projectId);
+    List<Task> findAllByProjectId(int projectId);
 }
